feat: 스터디 관리 페이지에서 스터디 선택 후 멤버를 관리 할 수 있도록 수정
