+++ src/App.jsx
@@ -3,15 +3,15 @@ import StudyDetail from "./pages/StudyDetail.jsx";
 import DetailedSearch from "./pages/DetailedSearch.jsx";
 import StudyCreate from './pages/StudyCreate';
 import CourseReviewForm from "./component/course/Course";
-import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
 import Login from "./pages/Login";
-import StudyMember from "./pages/StudyMember.jsx";
+import StudyManagement from "./pages/StudyManagement.jsx";
 
 function App() {
   return (
     <Routes>
-      <Route path="/study-member" element={<StudyMember />} />
+      <Route path="/study-manage" element={<StudyManagement />} />
       <Route path="/" element={<Home />} />
       <Route path="/login" element={<Login />} />
       <Route path="/list" element={<List />} />
